adapt to x64

+++ BigHomework/DivideWords.cpp
@@ -62,7 +62,7 @@ Stack<CharString> getInfoFileList() {
     CharString path;
     path = L".\\output\\*.info";
 
-    long hFile = 0;
+    intptr_t hFile = 0;
     struct _wfinddata_t fileData;
     const wchar_t * dirPath = path.wchar();
     if ((hFile = _wfindfirst(dirPath, &fileData)) != -1) {
